Revert "test: create an error intentionally to test if web integration test CI work correctly"
This reverts commit 52af4eed3faf178d647ff7e1b7024392009df534.

+++ example/integration_test/all_tests.dart
@@ -52,7 +52,7 @@ void main() {
         final shape = [2, 2]; // 2x2 matrix
 
         final tensor = await OrtValue.fromList(inputData, shape);
-        expect(tensor.dataType, OrtDataType.int32);
+        expect(tensor.dataType, OrtDataType.float32);
         expect(tensor.shape, shape);
 
         final retrievedData = await tensor.asFlattenedList();
